+++ test/html_parser_test.dart
@@ -12,16 +12,14 @@ void main() {
   testWidgets('Tests some plain old text', (WidgetTester tester) async {
     await tester.pumpWidget(MaterialApp(
         home: Scaffold(
-          body: Html(data: "This is some plain text"),
-        )
-    ));
+      body: Html(data: "This is some plain text"),
+    )));
 
     expect(find.text("This is some plain text"), findsOneWidget);
-
   });
 
-  testWidgets('Tests that a <b> element gets rendered correctly', (WidgetTester tester) async {
-
+  testWidgets('Tests that a <b> element gets rendered correctly',
+      (WidgetTester tester) async {
     await tester.pumpWidget(MaterialApp(
       home: Scaffold(
         body: Html(
@@ -31,11 +29,11 @@ void main() {
     ));
 
     expect(find.byType(RichText), findsOneWidget);
-
   });
 
-  testWidgets('Tests that a combination of elements and text nodes gets rendered', (WidgetTester tester) async {
-
+  testWidgets(
+      'Tests that a combination of elements and text nodes gets rendered',
+      (WidgetTester tester) async {
     await tester.pumpWidget(MaterialApp(
       home: Scaffold(
         body: Html(
@@ -46,11 +44,10 @@ void main() {
 
     expect(find.byType(RichText), findsNWidgets(2));
     expect(find.text(" and plain text"), findsOneWidget);
-
   });
 
-  testWidgets('Tests that a <i> element gets rendered correctly', (WidgetTester tester) async {
-
+  testWidgets('Tests that a <i> element gets rendered correctly',
+      (WidgetTester tester) async {
     await tester.pumpWidget(MaterialApp(
       home: Scaffold(
         body: Html(
@@ -62,39 +59,37 @@ void main() {
     expect(find.byType(RichText), findsNWidgets(4));
     expect(find.text(", "), findsOneWidget);
     expect(find.text(" and plain text"), findsOneWidget);
-
   });
 
-  testWidgets('Tests that nested elements get rendered correctly', (WidgetTester tester) async {
-
+  testWidgets('Tests that nested elements get rendered correctly',
+      (WidgetTester tester) async {
     await tester.pumpWidget(MaterialApp(
       home: Scaffold(
         body: Html(
-          data: "<b>Bold Text and <i>Italic bold text and <u>Underlined italic bold text</u></i></b>",
+          data:
+              "<b>Bold Text and <i>Italic bold text and <u>Underlined italic bold text</u></i></b>",
         ),
       ),
     ));
 
     expect(find.byType(RichText), findsOneWidget);
-
   });
 
-  testWidgets('Tests that the header elements (h1-h6) get rendered correctly', (WidgetTester tester) async {
-
+  testWidgets('Tests that the header elements (h1-h6) get rendered correctly',
+      (WidgetTester tester) async {
     await tester.pumpWidget(MaterialApp(
       home: Scaffold(
         body: Html(
-          data: "<h1>H1</h1><h2>H2</h2><h3>H3</h3><h4>H4</h4><h5>H5</h5><h6>H6</h6>",
+          data:
+              "<h1>H1</h1><h2>H2</h2><h3>H3</h3><h4>H4</h4><h5>H5</h5><h6>H6</h6>",
         ),
       ),
     ));
 
     expect(find.byType(RichText), findsNWidgets(6));
-
   });
 
   testWidgets('Tests the provided example', (WidgetTester tester) async {
-
     await tester.pumpWidget(MaterialApp(
       home: Scaffold(
         body: Html(
@@ -123,6 +118,5 @@ void main() {
 
     //Expect 3. One created by Html widget as part of the container, one for the <body>, and one for the <div>
     expect(find.byType(Column), findsNWidgets(3));
-
   });
 }